Remove Hansol** files and merge into original files
