+++ test/hardware_buttons_test.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:hardware_buttons/hardware_buttons.dart';
 
 void main() {
   const MethodChannel channel = MethodChannel('hardware_buttons');
@@ -15,7 +14,7 @@ void main() {
     channel.setMockMethodCallHandler(null);
   });
 
-  test('getPlatformVersion', () async {
-    expect(await HardwareButtons.platformVersion, '42');
-  });
+//  test('getPlatformVersion', () async {
+//    expect(await HardwareButtons.platformVersion, '42');
+//  });
 }
